Created separate file for http handling
